creating bugfix branch

+++ etc/release.sh
@@ -32,9 +32,10 @@ SVN_REPO_ROOT=http://seagull.phpkitchen.com:8172/svn/seagull
 ##############################
 function usage()
 {
-      echo "Usage: ./release.sh VERSION RELEASE"
-      echo "  where VERSION is the $PROJECT_NAME version (e.g. 0.4.4)"
-      echo "  and RELEASE is the release (e.g. release_0_4_4)"
+      echo ""
+      echo "Usage: ./release.sh version release password"
+      echo "    where \"version\" is the $PROJECT_NAME version (e.g. 0.3.1)"
+      echo "    and \"release\" is the release (e.g. release_0_3_1)"
 }
 
 ##############################
@@ -54,18 +55,29 @@ function checkArgs()
 ##############################
 function checkPreviousVersions()
 {
-    # Remove unuseful old stuff
-    echo "Removing last seagull export ..."
-    rm -rf /tmp/seagull-$VERSION
+    # Check that the release directory doesn't already exist:
+    if [ -d "/tmp/seagull-$VERSION" ]; then
+      echo "Removing last seagull export ..."
+      rm -rf /tmp/seagull-$VERSION
+    fi
     
-    echo "Removing last seagull tarball ..."
-    rm -f /tmp/seagull-$VERSION.tar.gz
+    # Check that the last tarball doesn't exist:
+    if [ -e "/tmp/seagull-$VERSION.tar.gz" ]; then
+      echo "Removing last seagull tarball ..."
+      rm -f /tmp/seagull-$VERSION.tar.gz
+    fi
 
-    echo "Removing last seagull apiDocs dir ..."
-    rm -rf /tmp/seagullApiDocs-$VERSION
+    # Check that the last apiDocs dir doesn't exist:
+    if [ -d "/tmp/seagullApiDocs-$VERSION" ]; then
+      echo "Removing last seagull apiDocs dir ..."
+      rm -rf /tmp/seagullApiDocs-$VERSION
+    fi
     
-    echo "Removing last seagull apiDocs tarball ..."
-    rm -f /tmp/seagullApiDocs-$VERSION.tar.gz
+    # Check that the last apiDocs tarball doesn't exist:
+    if [ -e "/tmp/seagullApiDocs-$VERSION.tar.gz" ]; then
+      echo "Removing last seagull apiDocs tarball ..."
+      rm -f /tmp/seagullApiDocs-$VERSION.tar.gz
+    fi
 }
 
 ##############################
@@ -73,6 +85,7 @@ function checkPreviousVersions()
 ##############################
 function tagRelease()
 {
+    # tag release
     $SVN copy $SVN_REPO_ROOT/trunk $SVN_REPO_ROOT/tags/$RELEASE
 }
 
@@ -81,6 +94,7 @@ function tagRelease()
 ##############################
 function exportSvnAndPackage()
 {   
+    # export release
     $SVN export $SVN_REPO_ROOT/trunk -r $RELEASE 
     
     #rename trunk to project name
@@ -111,8 +125,8 @@ function exportSvnAndPackage()
     mv $PROJECT_NAME $PROJECT_NAME-$VERSION
     
     # tar and zip
-    tar cvf $PROJECT_NAME-$VERSION.tar $PROJECT_NAME-$VERSION \
-        | gzip -f > $PROJECT_NAME-$VERSION.tar.gz
+    tar cvf $PROJECT_NAME-$VERSION.tar $PROJECT_NAME-$VERSION
+    gzip -f $PROJECT_NAME-$VERSION.tar
 }
 
 ##############################
@@ -143,6 +157,7 @@ function generateApiDocs()
     #execute phpDoc
     $PROJECT_NAME-$VERSION/etc/phpDocCli.sh
 
+    # rename folder    
     mv seagullApiDocs seagullApiDocs-$VERSION
 }
 
@@ -151,8 +166,8 @@ function generateApiDocs()
 ##############################
 function packageApiDocs()
 {
-    tar cvf seagullApiDocs-$VERSION.tar seagullApiDocs-$VERSION \
-        | gzip -f > seagullApiDocs-$VERSION.tar.gz
+    tar cvf seagullApiDocs-$VERSION.tar seagullApiDocs-$VERSION
+    gzip -f seagullApiDocs-$VERSION.tar
 }
 
 ##############################
@@ -177,8 +192,7 @@ EOF
 ##############################
 function scpChangelogToSglSite()
 {
-    scp -1 $PROJECT_NAME-$VERSION/CHANGELOG.txt \
-        demian@phpkitchen.com:/var/www/html/seagull/web/
+    scp -1 $PROJECT_NAME-$VERSION/CHANGELOG.txt demian@phpkitchen.com:/var/www/html/seagull/web/
 }
 
 ##############################
@@ -193,6 +207,7 @@ checkPreviousVersions
 
 #tagRelease
 
+# move to tmp dir
 cd /tmp
 
 exportSvnAndPackage
